refactor: Move Task based structs back into tasks.
Looks like we could just move DecoratedCallable up rather than bring everything else down.
Makes things slightly cleaner if we're not really providing this behaviour for editor agents
