+++ encord_agents/tasks/dependencies.py
@@ -7,11 +7,9 @@ import numpy as np
 from encord.exceptions import AuthenticationError, AuthorisationError, UnknownException
 from encord.objects.ontology_labels_impl import LabelRowV2
 from encord.orm.storage import StorageItemType
-from encord.project import Project
 from encord.storage import StorageItem
 from encord.user_client import EncordUserClient
 from encord.workflow.common import WorkflowTask
-from encord.workflow.stages.agent import AgentTask
 from encord.workflow.workflow import WorkflowStage
 from numpy.typing import NDArray
 from typing_extensions import Annotated
