+++ encord_agents/tasks/models.py
@@ -1,6 +1,26 @@
+from dataclasses import dataclass
+from typing import Callable, TypeVar
 from uuid import UUID
 
+from encord.objects.ontology_labels_impl import LabelRowV2
+from encord.storage import StorageItem
 from pydantic import BaseModel, Field
+
+TaskAgentReturnPathway = str | UUID | None
+
+
+@dataclass
+class TaskAgentReturnStruct:
+    """Return this from your agent and we will handle propagating the updates in batches"""
+
+    pathway: TaskAgentReturnPathway = None
+    storage_item: StorageItem | None = None
+    label_row: LabelRowV2 | None = None
+
+
+TaskAgentReturnType = TaskAgentReturnPathway | TaskAgentReturnStruct
+
+DecoratedCallable = TypeVar("DecoratedCallable", bound=Callable[..., TaskAgentReturnType])
 
 
 class AgentTaskConfig(BaseModel):
